requirements.txt
# AI / NLP
sentence-transformers
ollama>=0.1.8

# PDF / OCR / Docs
PyMuPDF
pytesseract
Pillow
python-docx

# Backend
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.6

# Database
motor>=3.4
pymongo>=4.6

# Env
python-dotenv>=1.0

requests>=2.31
